new file for stats tests
